move to the next screen on enter when writing room or player name
Treat InputWithButton as a form so that onClick is called on submit. That means pressing enter will now move you to the next screen.

InputName was modified so that the user joins as a player when enter is pressed. I briefly considered having them join as what they joined as last time, but that would probably be confusing due to its inconsistency.